refactor: 폴더구조 약간의 정리, auth 활성화
