+++ src/pages/WordSetListPage/components/WordSetItem.styled.ts
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { Colors } from "../../designs/colors";
+import { Colors } from "../../../designs/colors";
 
 const S = {
     Root: styled.div`
