+++ src/routes/_authenticated.tsx
@@ -8,10 +8,9 @@ export const Route = createFileRoute("/_authenticated")({
 function RouteComponent() {
     const { isLoginned } = useAuth();
 
-    // TODO: Uncomment this code to enable authentication
-    // if (!isLoginned) {
-    //     return <Navigate to="/login" />;
-    // }
+    if (!isLoginned) {
+        return <Navigate to="/login" />;
+    }
 
     return <Outlet />;
 }
