+++ src/routes/_authenticated/words.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import WordsMain from '../../pages/WordSetListPage/WordsMain'
+import WordsMain from '../../pages/WordSetListPage/components/WordsMain'
 
 export const Route = createFileRoute('/_authenticated/words')({
   component: WordsMain,
